Extract a library crate from the binary crate
To allow the library crate to be used to generate links from other
scripts too :)
