+++ src/main.rs
@@ -1,40 +1,13 @@
 #![feature(plugin)]
 #![plugin(clippy)]
 
-
 extern crate clap;
 extern crate glob;
-extern crate url;
+extern crate prlink;
 
 use clap::{App, Arg};
-use std::error::Error;
-use std::fs::File;
-use std::io::prelude::*;
-use url::form_urlencoded::Serializer;
 use glob::glob;
 use std::path::PathBuf;
-
-
-fn read_file(path: &str) -> String {
-    let mut file = match File::open(&path) {
-        Err(why) => panic!("couldn't open {}: {}", path, Error::description(&why)),
-        Ok(file) => file,
-    };
-    let mut s = String::new();
-    match file.read_to_string(&mut s) {
-        Err(why) => panic!("couldn't read {}: {}", path, Error::description(&why)),
-        Ok(_) => s,
-    }
-}
-
-fn linkify_file(path: &PathBuf) -> String {
-    let data = &[("code".to_owned(), read_file(path.to_str().unwrap()))];
-    let s = Serializer::new(String::new())
-        .extend_pairs(data)
-        .finish();
-    let fname = path.file_name().unwrap().to_string_lossy();
-    format!("- [{:?}](https://play.rust-lang.org/?{})", fname, s)
-}
 
 fn main() {
     let matches = App::new("prlink")
@@ -55,7 +28,7 @@ fn main() {
     for entry in glob(absolute_path.to_str().unwrap()).unwrap() {
         match entry {
             Ok(path) => {
-                println!("{}", linkify_file(&path));
+                println!("{}", prlink::linkify_file(&path));
             },
             Err(e) => println!("{:?}", e),
         }
